ai-service/requirements.txt
torch
torchvision
fastapi
uvicorn
python-multipart
pillow
pandas
numpy
opencv-python